umbra: split precomputations of individual queries
